% Generated by roxygen2 (4.0.1): do not edit by hand
\name{check}
\alias{check}
\title{Check Package is Ready to Use}
\usage{
check(path = ".", error = TRUE, quiet = FALSE)
}
\arguments{
\item{path}{Path to the package}

\item{error}{Logical indicating if problems should be treated as errors}

\item{quiet}{Logical indicating if a description of problems
should be printed.}
}
\value{
An invisible logical value indicating if the package looks
ready for use with RcppR6.  However, if the package is not ready
and \code{error} is \code{TRUE}, then nothing is returned as the
function will throw an error.
}
\description{
Check that a package is ready for use with RcppR6.  This just
checks for our requirements and prints diagnostics.  It is
probably unsufficient, but hopefully provides enough information.
}
\author{
Rich FitzJohn
}

