% Generated by roxygen2 (4.0.1): do not edit by hand
\name{create}
\alias{create}
\title{Create New Package, With RcppR6 Support}
\usage{
create(path, ...)
}
\arguments{
\item{path}{Location of the package.  See
\code{\link[devtools]{create}} for more information.}

\item{...}{Additional arguments passed to
\code{\link[devtools]{create}}.}
}
\description{
Wrapper around \code{devtools::create} that also creates RcppR6
files.
}
\author{
Rich FitzJohn
}

