% Generated by roxygen2 (4.0.1): do not edit by hand
\name{package_name}
\alias{package_name}
\title{Determine name of a package}
\usage{
package_name(path = ".")
}
\arguments{
\item{path}{Path to the package (DESCRIPTION must be in this
directory).  Defaults to the current directory.}
}
\description{
Determine name of a package
}

