% Generated by roxygen2 (4.0.1): do not edit by hand
\name{path}
\alias{path}
\title{Path Name of Installed Package}
\usage{
path()
}
\value{
Nothing -- called for the side effect of printing the path
name to the console.
}
\description{
Internal function to find the location of the package.  This is
used in the Makefiles that it provides.
}

