% Generated by roxygen2 (4.0.1): do not edit by hand
\name{RcppR6}
\alias{RcppR6}
\alias{install}
\title{Update Or Install RcppR6 Files}
\usage{
RcppR6(path = ".", verbose = TRUE, install = FALSE, attributes = TRUE,
  roxygen = FALSE)

install(...)
}
\arguments{
\item{path}{Path to the package (this directory must contain the
DESCRIPTION file)}

\item{verbose}{Logical indicating if information about the process
will be generated.  It's not all that verbose really.}

\item{install}{Logical indicating if this should be treated as a
fresh install.  Specifying \code{TRUE} (not the default) should
always be safe, but will copy default or skeleton copyies of files
into place if they do not exist, as well as update your
DESCRIPTION file.}

\item{attributes}{Should Rcpp attributes be regenerated as well?
This is probably a good idea (and is the default).}

\item{roxygen}{Should \code{devtools::document} be run to rebuild
the NAMESPACE and documentation files?  This is generally a good
idea, but will force a rebuild of the source code in the package.
That might not always be desirable as you can end up in unloadable
states.}

\item{...}{Arguments passed to \code{RcppR6()}}
}
\description{
Update or install RcppR6 files.  This will copy required files
around, parse your \code{inst/classes.yml} file, and generate
required files.  Using \code{RcppR6::install()} is equivalent to
passing \code{install=TRUE} to \code{RcppR6::RcppR6}.
}
\details{
More details coming later!
}
\author{
Rich FitzJohn
}

