% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rcppr6_file}
\alias{rcppr6_file}
\title{Get RCPPR6 File}
\usage{
rcppr6_file(...)
}
\arguments{
\item{...}{Passed to \code{\link{system.file}}}
}
\description{
Extract a file from the rcppr6 (this!) package.  Just sets some
defaults to \code{\link{system.file}}
}

