% Generated by roxygen2 (4.0.1): do not edit by hand
\name{update_rcppr6}
\alias{install}
\alias{update_rcppr6}
\title{Update Or Install rcppr6 Files}
\usage{
update_rcppr6(path = ".", verbose = TRUE, install = FALSE)

install(...)
}
\arguments{
\item{path}{Path to the package (this directory must contain the
DESCRIPTION file)}

\item{verbose}{Logical indicating if information about the process
will be generated.  It's not all that verbose really.}

\item{install}{Logical indicating if this should be treated as a
fresh install.  Specifying \code{TRUE} (not the default) should
always be safe, but will copy default or skeleton copyies of files
into place if they do not exist, as well as update your
DESCRIPTION file.}

\item{...}{Arguments passed to \code{update_rcppr6}}
}
\description{
Update or install rcppr6 files.  This will copy required files
around, parse your \code{inst/classes.yml} file, and generate
required files.  Using \code{rcppr6::install()} is equivalent to
passing \code{install=TRUE} to \code{rcppr6::update_rcppr6}.
}
\details{
More details coming later!
}
\author{
Rich FitzJohn
}

