% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/api.R
\name{RcppR6}
\alias{RcppR6}
\alias{install}
\title{Update Or Install RcppR6 Files}
\usage{
RcppR6(path, install = FALSE, attributes = TRUE, verbose = TRUE)

install(...)
}
\arguments{
\item{path}{Path to the package (this directory must contain the
DESCRIPTION file)}

\item{install}{Logical indicating if this should be treated as a
fresh install.  Specifying \code{TRUE} (not the default) should
always be safe, but will copy default or skeleton copyies of files
into place if they do not exist, as well as update your
DESCRIPTION file.}

\item{attributes}{Should Rcpp attributes be regenerated as well?
This is probably a good idea (and is the default).}

\item{verbose}{Logical indicating if information about the process
will be generated.  It's not all that verbose really.}

\item{...}{Arguments passed to \code{RcppR6()}}
}
\description{
Update or install RcppR6 files.  This will copy required files
around, parse your \code{inst/classes.yml} file, and generate
required files.  Using \code{RcppR6::install()} is equivalent to
passing \code{install=TRUE} to \code{RcppR6::RcppR6}.
}
\details{
More details coming later!
}

