% Generated by roxygen2 (4.0.1): do not edit by hand
\name{RcppR6_file}
\alias{RcppR6_file}
\title{Get RcppR6 File}
\usage{
RcppR6_file(...)
}
\arguments{
\item{...}{Passed to \code{\link{system.file}}}
}
\description{
Extract a file from the RcppR6 (this!) package.  Just sets some
defaults to \code{\link{system.file}}
}

